Deprecate backlog_no_timestamp parameter
backlog_timestamp must be used instead.

The default behavior is unchanged and can explicitly used with:

    backlog_timestamp = "time";

The "none" value can be used with the backlog_timestamp parameter in
order to use the same behavior that the "no" value of the deprecated
parameter backlog_no_timestamp provided.

The date can now be displayed (in addition to the time) while backlogging:

    backlog_timestamp = "datetime";